# TM8/tm8-game-mobile-release-0.6.0/input_folder/dev/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/auth/token/refresh": {
            "post": {
                "operationId": "AuthController_refreshTokens",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/sign-in": {
            "post": {
                "operationId": "AuthController_login",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthLoginInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "operationId": "AuthController_register",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/verify-phone": {
            "post": {
                "operationId": "AuthController_veirifyPhone",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyCodeInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/resend-code": {
            "post": {
                "operationId": "AuthController_resendCode",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PhoneVerificationInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK response"
                    },
                    "204": {
                        "description": ""
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/google/verify": {
            "post": {
                "operationId": "AuthController_googleAuth",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyGoogleIdInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/epic-games/sign-in": {
            "get": {
                "operationId": "AuthController_epicGamesAuth",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK response"
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/epic-games/verify": {
            "post": {
                "operationId": "AuthController_epicGamesVerify",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EpicGamesVerifyInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK response"
                    },
                    "201": {
                        "description": ""
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/epic-games": {
            "delete": {
                "operationId": "AuthController_removeEpicAccoun",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "OK response"
                    },
                    "204": {
                        "description": ""
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/forgot-password": {
            "post": {
                "operationId": "AuthController_forgotPassword",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ForgotPasswordInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK response"
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/forgot-password/verify": {
            "post": {
                "operationId": "AuthController_verifyForgotPassword",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyCodeInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK response"
                    },
                    "201": {
                        "description": ""
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/forgot-password/reset": {
            "post": {
                "operationId": "AuthController_resetPassword",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/date-of-birth": {
            "patch": {
                "operationId": "AuthController_setDateOfBirth",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SetDateOfBirthInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK response"
                    },
                    "204": {
                        "description": ""
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/set-phone-number": {
            "patch": {
                "operationId": "AuthController_setUserPhone",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SetUserPhoneInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
        "/api/v1/auth/apple/verify": {
            "post": {
                "operationId": "AuthController_verifyAppleIdToken",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VerifyAppleIdInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Non-2XX response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TGMDExceptionResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Auth"
                ]
            }
        },
...
    },
    "info": {
        "title": "TM8 API",
        "description": "TM8 API",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "TGMDExceptionResponse": {
                "type": "object",
                "properties": {
                    "origin": {
                        "enum": [
                            "HTTP",
                            "Mongo",
                            "TGMD",
                            "Unknown"
                        ],
                        "type": "string"
                    },
                    "status": {
                        "type": "number",
                        "description": "HTTP-compatible status code"
                    },
                    "code": {
                        "type": "string"
                    },
                    "exception": {
                        "type": "string",
                        "description": "Exception name"
                    },
                    "detail": {
                        "type": "string",
                        "description": "Human-readable error message"
                    }
                },
                "required": [
                    "origin",
                    "status",
                    "code",
                    "exception",
                    "detail"
                ]
            },
...
            "AuthResponse": {
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string"
                    },
                    "phoneVerificationRequired": {
                        "type": "boolean"
                    },
                    "refreshToken": {
                        "type": "string"
                    },
                    "success": {
                        "type": "boolean"
                    },
                    "username": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "signupType": {
                        "type": "string"
                    },
                    "chatToken": {
                        "type": "string"
                    },
                    "dateOfBirth": {
                        "format": "date-time",
                        "type": "string",
                        "nullable": true
                    }
                },
                "required": [
                    "success"
                ]
            },
            "AuthLoginInput": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "notificationToken": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "RegisterInput": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "dateOfBirth": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "timezone": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "notificationToken": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "username",
                    "dateOfBirth",
                    "password"
                ]
            },
            "VerifyCodeInput": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "number"
                    }
                },
                "required": [
                    "code"
                ]
            },
            "PhoneVerificationInput": {
                "type": "object",
                "properties": {
                    "phoneNumber": {
                        "type": "string"
                    }
                },
                "required": [
                    "phoneNumber"
                ]
            },
            "VerifyGoogleIdInput": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "fullName": {
                        "type": "string"
                    },
                    "notificationToken": {
                        "type": "string"
                    }
                },
                "required": [
                    "token"
                ]
            },
            "EpicGamesVerifyInput": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "string"
                    }
                },
                "required": [
                    "code",
                    "userId"
                ]
            },
            "ForgotPasswordInput": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    }
                }
            },
...
            "SetDateOfBirthInput": {
                "type": "object",
                "properties": {
                    "dateOfBirth": {
                        "type": "string"
                    }
                },
                "required": [
                    "dateOfBirth"
                ]
            },
...
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}